% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_NB.R
\name{alpha_NB}
\alias{alpha_NB}
\title{Obtain alpha for a Poisson offspring distribution}
\usage{
alpha_NB(R, over)
}
\arguments{
\item{R}{is the reproduction number, i.e. the average number of secondary cases due to a single case.
This can be any positive number. if R is a vector, then the length of R must be 'n' (see below).}

\item{over}{is the overdispersion in the offspring distribution.}
}
\value{
The function returns alpha assuming a NB offspring distribution.
}
\description{
Same as alpha_Poisson but applicable for NB offspring distribution with overdispersion 'over'
}
\examples{

x <- alpha_NB(R = 1.5, over = 1)
x

}
\author{
Pierre Nouvellet (\email{p.nouvellet@imperial.ac.uk})
}
