% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_poisson.R
\name{alpha_poisson}
\alias{alpha_poisson}
\title{Obtain alpha for a Poisson offspring distribution}
\usage{
alpha_poisson(R)
}
\arguments{
\item{R}{is the reproduction number, i.e. the average number of secondary cases due to a single case.
This can be any positive number. if R is a vector, then the length of R must be 'n' (see below).}
}
\value{
The function returns alpha assuming a Poisson offspring distribution.
}
\description{
If the transmission can be charaterised by a Poisson offspring distribution, the function
alpha.poisson(R) compute alpha for a given R.
alpha is linked to probability of extinction and allow us, for any Rl>1, to find an Rs<1 guaranteeing that
conditional on extinction, outbreak properties starting from Rs or Rl are exactly the same.
}
\examples{

x <- alpha_poisson(c(1.5,.5))
x

}
\author{
Pierre Nouvellet (\email{p.nouvellet@imperial.ac.uk})
}
