% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_Lhood_NB.R
\name{element_Lhood_NB}
\alias{element_Lhood_NB}
\title{# Likelihood of observing outbreaks of size y given parameters}
\usage{
element_Lhood_NB(R, over, z, g, g0)
}
\arguments{
\item{R}{is the reproduction number, i.e. the average number of secondary cases due to a single case.
This can be any positive number.}

\item{over}{is the overdispersion in the offspring distribution.}

\item{z}{is the true potential outbreak sizes. It is precomputed by proba_observation.}

\item{g:}{the probability of observing size y given z,rho; with z the true number of
cases (reported and un-reported cases) and rho the reporting probability. It is precomputed by proba_observation.}

\item{g0:}{the probability of not observing an outbreak (i.e. y=0 or no cases) given z and rho.
It is precomputed by proba_observation.}
}
\value{
The function returns the likelihood of the observations for a given {R,rho,overdispersion}.
}
\description{
The function internally relies on pre-computing:
 1) g: the probability of observing size y given z,rho; with z the true number of
 cases (reported and un-reported cases) and rho the reporting probability
}
\details{
2) g0: the probability of not observing an outbreak (i.e. y=0 or no cases) given z and rho.

g and g0 are obtained using the function proba_observation.

adapted from element_Lhood_poisson but accounting for a NB offspring distribution with overdispersion 'over'
}
\examples{

y <- element_Lhood_NB(R = .5, over = 1, z = x$possible_size, g = x$p_y_z, g0 = x$p_0_z)
y

}
\author{
Pierre Nouvellet (\email{p.nouvellet@imperial.ac.uk})
}
