% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_z_NB.R
\name{f_z_NB}
\alias{f_z_NB}
\title{Compute the probability of observing an outbreak of size z given an R<1 and assuming a NB offspring distribution}
\usage{
f_z_NB(z, s, R, over)
}
\arguments{
\item{R}{is the reproduction number, i.e. the average number of secondary cases due to a single case.
This can be any positive number.}

\item{over}{is the overdispersion in the offspring distribution.}
}
\value{
The function returns a list including:
\itemize{

\item the probability of observing an outbreak of size z given R/overdispersion

}
}
\description{
adapted from f_z_poisson but accounting for a NB offspring distribution with overdispersion 'over'
}
\examples{

x <- f_z_NB(z = 1:1e2, s = 1, R = .5, over = 1)


}
\author{
Pierre Nouvellet (\email{p.nouvellet@imperial.ac.uk})
}
