% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{# Likelihood of observing outbreaks of size y given parameters}
\usage{
import(y_obs, profile, threshold_z, threshold_import, CI)
}
\arguments{
\item{R}{is the reproduction number, i.e. the average number of secondary cases due to a single case.
This can be any positive number.}

\item{z}{is the true potential outbreak sizes. It is precomputed by proba_observation.}

\item{g:}{the probability of observing size y given z,rho; with z the true number of
cases (reported and un-reported cases) and rho the reporting probability. It is precomputed by proba_observation.}

\item{g0:}{the probability of not observing an outbreak (i.e. y=0 or no cases) given z and rho.
It is precomputed by proba_observation.}
}
\value{
The function returns the likelihood of the observations for a given {R,rho}.
}
\description{
The function internally relies on pre-computing:
 1) g: the probability of observing size y given z,rho; with z the true number of
 cases (reported and un-reported cases) and rho the reporting probability
}
\details{
2) g0: the probability of not observing an outbreak (i.e. y=0 or no cases) given z and rho.

g and g0 are obtained using the function proba_observation.
}
\examples{

x <- element_Lhood_poisson(.5,x$possible_size,x$p_y_z,x$p_0_z)
x

}
\author{
Pierre Nouvellet (\email{p.nouvellet@imperial.ac.uk})
}
