% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba_ext.R
\name{proba_ext}
\alias{proba_ext}
\title{Compute the probability of extinction for a given alpha}
\usage{
proba_ext(alpha)
}
\arguments{
\item{alpha:}{a value related to the reproduction number and the offspring distribution.}
}
\value{
The function returns the probability of extinction for a given alpha. Formally, Pext=exp(-alpha)
}
\description{
If the transmission can be charaterised by a Poisson offspring distribution, see the function
alpha.poisson(R) to compute alpha for a given R.
The probability of extinction is computed assuming a single initial case. To obtain the probability
extinction associated with n initial case, simply raise the probability returned by this function to the power of n.
}
\examples{

x <- proba.ext(c(.5,1))
x

}
\author{
Pierre Nouvellet (\email{p.nouvellet@imperial.ac.uk})
}
