% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_eff_NB.R
\name{R_eff_NB}
\alias{R_eff_NB}
\title{Obtain an effective R when conditioning on extinction and assuming poisson}
\usage{
R_eff_NB(R, over)
}
\arguments{
\item{R}{is the reproduction number, i.e. the average number of secondary cases due to a single case.
This can be any positive number. if R is a vector, then the length of R must be 'n' (see below).}

\item{over}{is the overdispersion in the offspring distribution.}
}
\value{
The function returns a list including:
\itemize{

\item R_eff: the effective reproduction number. Note that if R<1, then R_eff=R.

\item alpha given the reproduction number and assuming a Poisson offspring distribution. Note that if R<1, then alpha=0.

\item proba_ext: the probability of extinction given alpha. Note that if R<1, then proba_ext=1.

}
#'
}
\description{
For any Rl>1, we can find an Rs<1 guaranteeing that
conditional on extinction, outbreak properties starting from Rs or Rl are exactly the same.
We call this Rs an effective R.
The func`tion compute the effetive R as well as alpha (see alpha_poisson) and the probability of extinction (see proba_ext)
}
\details{
adapted from R_eff_poisson but accounting for a NB offspring distribution with overdispersion 'over'
}
\examples{

x <- R_eff_NB(R = 1.5, over = 1)
x

}
\author{
Pierre Nouvellet (\email{p.nouvellet@imperial.ac.uk})
}
