% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_eff_poisson.R
\name{R_eff_poisson}
\alias{R_eff_poisson}
\title{Obtain an effective R when conditionning on extinction and assuming poisson}
\usage{
R_eff_poisson(R)
}
\arguments{
\item{R}{is the reproduction number, i.e. the average number of secondary cases due to a single case.
This can be any positive number. if R is a vector, then the length of R must be 'n' (see below).}
}
\value{
The function returns a list including:
\itemize{

\item R_eff: the effective reproduction number. Note that if R<1, then R_eff=R.

\item alpha given the reproduction number and assuming a Poisson offspring distribution. Note that if R<1, then alpha=0.

\item proba_ext: the probability of extinction given alpha. Note that if R<1, then proba_ext=1.

}
#'
}
\description{
For any Rl>1, we can find an Rs<1 guaranteeing that
conditional on extinction, outbreak properties starting from Rs or Rl are exactly the same.
We call this Rs an effective R.
The funtion compute the effetive R as well as alpha (see alpha_poisson) and the probability of extinction (see proba_ext)
}
\examples{

x <- R_eff_poisson(.5)
x
x <- R_eff_poisson(1.5)
x

}
\author{
Pierre Nouvellet (\email{p.nouvellet@imperial.ac.uk})
}
