% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_max_likelihood.R
\name{theta_max_likelihood}
\alias{theta_max_likelihood}
\title{Obatin the maximium likleihood from a vector of parameter and its likelihood profile}
\usage{
theta_max_likelihood(theta, likelihood, threshold_CI)
}
\arguments{
\item{theta}{is a vector containing parameter's values for which the likelihood has been evaluated}

\item{likelihood}{at the paraneter's value (log-likelihood)}

\item{threshold_CI}{is the threshold to estimate the confidence interval, i.e. to obtain
a 95% confidence interval threshold_CI=95.}
}
\value{
The function returns a list including:
\itemize{

\item theta_max_likelihood: the maximum likelihood estimated with the rabge of theta values evaluated.

\item max_likelihood: the log of the maximum likelihood.

\item lower_theta,upper_theta: lower and upper bound of the 'threshold_CI'% confidence interval.


}
}
\description{
Also gives confidence interval
}
\author{
Pierre Nouvellet (\email{p.nouvellet@imperial.ac.uk})
}
