% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_z_poiss.R
\name{f_z_poiss}
\alias{f_z_poiss}
\title{Compute the probability of observing an outbreak of size z given an R<1 and assuming a Poisson offspring distribution}
\usage{
f_z_poiss(z, s, R)
}
\arguments{
\item{R}{is the reproduction number, i.e. the average number of secondary cases due to a single case.
This can be any positive number.}
}
\value{
The function returns a list including:
\itemize{

 \item the probability of observing an outbreak of size z given R.
 
 }
}
\description{
Compute the probability of observing an outbreak of size z given an R<1 and assuming a Poisson offspring distribution
}
\examples{

x <- f_z_poiss(z = 1:1e2, s = 1, R = .5)
x

}
\author{
Pierre Nouvellet (\email{p.nouvellet@imperial.ac.uk})
}
