% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_likelihood.R
\name{profile_likelihood}
\alias{profile_likelihood}
\title{Profile likelihood for R given observations, rho, threshold, accuracy maximum value for R}
\usage{
profile_likelihood(y_obs, rho, accuracy, max_R, over = NULL)
}
\arguments{
\item{over}{is the overdispersion in the offspring distribution.}

\item{y_obs:}{a vector of outbreak sizes observations.}

\item{rho:}{the probability of a single case being detected.}

\item{accuracy:}{R will take potential values in a grid spaning from accuracy up to max_R with increments 'accuracy'}

\item{max_R:}{R will take potential values in a grid spaning from accuracy up to max_R with increments 'accuracy'}
}
\value{
The function returns a list including:
\itemize{

\item theta: a vector of potential parameter values for which the likelihood has been evaluated.

\item Likelihood: a vector of likelihood evaluated for the potential R values.

}
#'
}
\description{
Profile likelihood for R given observations, rho, threshold, accuracy maximum value for R
}
\examples{

x <- profile_likelihood(c(1,2,3),.5,1e3,0.01,10)
x


}
\author{
Pierre Nouvellet (\email{p.nouvellet@imperial.ac.uk})
}
