% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_poisson.R
\name{simulate_poisson}
\alias{simulate_poisson}
\title{Simulate bla bla (title)}
\usage{
simulate_poisson(x, y)
}
\arguments{
\item{x}{this is what x is}

\item{y}{this is what y is}
}
\value{
The function returns ....
}
\description{
This is what it does bla bla (descripion). This is \code{some code}.
Bullet list:
}
\details{
\itemize{

\item toto: il est beau

\item tata: il est caca

}
}
\examples{

x <- simulate_poisson(1, 2)
x

}
\author{
Pierre Nouvellet (\email{p.nouvellet@imperial.ac.uk})
}
